+++ src/codec/i64-array.ts
@@ -1,8 +1,15 @@
 import type { ByteCursor } from "../core/index.js"
+import { DEV, assert } from "../util/assert.js"
 import { I64_BYTE_COUNT } from "../util/constants.js"
 import { IS_LITTLE_ENDIAN_PLATFORM } from "../util/util.js"
+import { isU32 } from "../util/validator.js"
 import { readFixedData } from "./data.js"
-import { readI64, readUintSafe, writeI64, writeUintSafe } from "./primitive.js"
+import {
+    readI64,
+    readUintSafe32,
+    writeI64,
+    writeUintSafe32,
+} from "./primitive.js"
 import { writeU8FixedArray } from "./u8-array.js"
 
 export const readI64FixedArray = IS_LITTLE_ENDIAN_PLATFORM
@@ -10,15 +17,21 @@ export const readI64FixedArray = IS_LITTLE_ENDIAN_PLATFORM
     : readI64FixedArrayBE
 
 export function readI64Array(bc: ByteCursor): BigInt64Array {
-    return readI64FixedArray(bc, readUintSafe(bc))
+    return readI64FixedArray(bc, readUintSafe32(bc))
 }
 
 function readI64FixedArrayLE(bc: ByteCursor, len: number): BigInt64Array {
+    if (DEV) {
+        assert(isU32(len))
+    }
     const byteCount = len * I64_BYTE_COUNT
     return new BigInt64Array(readFixedData(bc, byteCount))
 }
 
 function readI64FixedArrayBE(bc: ByteCursor, len: number): BigInt64Array {
+    if (DEV) {
+        assert(isU32(len))
+    }
     bc.check(len * I64_BYTE_COUNT)
     const result = new BigInt64Array(len)
     for (let i = 0; i < len; i++) {
@@ -32,7 +45,7 @@ export const writeI64FixedArray = IS_LITTLE_ENDIAN_PLATFORM
     : writeI64FixedArrayBE
 
 export function writeI64Array(bc: ByteCursor, x: BigInt64Array): void {
-    writeUintSafe(bc, x.length)
+    writeUintSafe32(bc, x.length)
     if (x.length !== 0) {
         writeI64FixedArray(bc, x)
     }