+++ src/codec/u32-array.ts
@@ -1,8 +1,15 @@
 import type { ByteCursor } from "../core/index.js"
+import { DEV, assert } from "../util/assert.js"
 import { U32_BYTE_COUNT } from "../util/constants.js"
 import { IS_LITTLE_ENDIAN_PLATFORM } from "../util/util.js"
+import { isU32 } from "../util/validator.js"
 import { readFixedData } from "./data.js"
-import { readU32, readUintSafe, writeU32, writeUintSafe } from "./primitive.js"
+import {
+    readU32,
+    readUintSafe32,
+    writeU32,
+    writeUintSafe32,
+} from "./primitive.js"
 import { writeU8FixedArray } from "./u8-array.js"
 
 export const readU32FixedArray = IS_LITTLE_ENDIAN_PLATFORM
@@ -10,15 +17,21 @@ export const readU32FixedArray = IS_LITTLE_ENDIAN_PLATFORM
     : readU32FixedArrayBE
 
 export function readU32Array(bc: ByteCursor): Uint32Array {
-    return readU32FixedArray(bc, readUintSafe(bc))
+    return readU32FixedArray(bc, readUintSafe32(bc))
 }
 
 function readU32FixedArrayLE(bc: ByteCursor, len: number): Uint32Array {
+    if (DEV) {
+        assert(isU32(len))
+    }
     const byteCount = len * U32_BYTE_COUNT
     return new Uint32Array(readFixedData(bc, byteCount))
 }
 
 function readU32FixedArrayBE(bc: ByteCursor, len: number): Uint32Array {
+    if (DEV) {
+        assert(isU32(len))
+    }
     bc.check(len * U32_BYTE_COUNT)
     const result = new Uint32Array(len)
     for (let i = 0; i < len; i++) {
@@ -32,7 +45,7 @@ export const writeU32FixedArray = IS_LITTLE_ENDIAN_PLATFORM
     : writeU32FixedArrayBE
 
 export function writeU32Array(bc: ByteCursor, x: Uint32Array): void {
-    writeUintSafe(bc, x.length)
+    writeUintSafe32(bc, x.length)
     if (x.length !== 0) {
         writeU32FixedArray(bc, x)
     }