+++ src/util/assert.ts
@@ -1,3 +1,5 @@
+export { DEV } from "#dev"
+
 export class AssertionError extends Error {
     constructor(message: string) {
         super(message)